+++ CanonizerForgotPasswordPage.py
